+++ lib/src/data/models/vietmap_boundaries_model.dart
@@ -1,8 +1,19 @@
+/// VietmapBoundaries is a model class that is used to parse the boundaries information from the server.
+/// This describes the administrative boundaries such as ward, district, and city.
 class VietmapBoundaries {
+  /// Type of boundary (0 for city, 1 for district, 2 for ward)
   final num? type;
+
+  /// Unique identifier for the boundary
   final num? id;
+
+  /// Name of the boundary
   final String? name;
+
+  /// Prefix of the boundary (e.g., "Phường" for ward)
   final String? prefix;
+
+  /// Full name of the boundary (e.g., "Phường 9" for ward)
   final String? fullName;
 
   VietmapBoundaries(