+++ lib/src/di/vietmap_context.dart
@@ -1,7 +1,16 @@
+import 'package:vietmap_flutter_plugin/src/core/enums/tilemap_enum.dart';
+import 'package:vietmap_flutter_plugin/src/domain/entities/vietmap_autocomplete_params_v4.dart';
+import 'package:vietmap_flutter_plugin/src/domain/entities/vietmap_migrate_address_params.dart';
+import 'package:vietmap_flutter_plugin/src/domain/entities/vietmap_reverse_params.dart';
 import 'package:vietmap_flutter_plugin/src/domain/repository/vietmap_api_repositories.dart';
+import 'package:vietmap_flutter_plugin/src/domain/usecase/geocode_v4_usecase.dart';
 import 'package:vietmap_flutter_plugin/src/domain/usecase/geocode_usecase.dart';
 import 'package:vietmap_flutter_plugin/src/domain/usecase/get_location_from_latlng_usecase.dart';
+import 'package:vietmap_flutter_plugin/src/domain/usecase/get_place_detail_v4_usecase.dart';
 import 'package:vietmap_flutter_plugin/src/domain/usecase/matrix_usecase.dart';
+import 'package:vietmap_flutter_plugin/src/domain/usecase/migrate_address_usecase.dart';
+import 'package:vietmap_flutter_plugin/src/domain/usecase/get_location_from_latlng_v4_usecase.dart';
+import 'package:vietmap_flutter_plugin/src/domain/usecase/search_address_v4_usecase.dart';
 
 import '../../vietmap_flutter_plugin.dart';
 import '../domain/usecase/get_direction_usecase.dart';
@@ -38,6 +47,29 @@ class Vietmap {
     return 'https://maps.vietmap.vn/api/maps/light/styles.json?apikey=$_vietmapAPIKey';
   }
 
+  /// Get vietmap style url based on TileMapEnum. See [TileMapEnum] for more details.
+  static String getMapStyle({TileMapEnum? type}) {
+    if (_vietmapAPIKey.isEmpty) {
+      throw Exception('Please call `Vietmap.getInstance(apiKey)` before use');
+    }
+    switch (type) {
+      case TileMapEnum.vectorDefault:
+        return 'https://maps.vietmap.vn/maps/styles/tm/style.json?apikey=$_vietmapAPIKey';
+      case TileMapEnum.vectorLight:
+        return 'https://maps.vietmap.vn/maps/styles/lm/style.json?apikey=$_vietmapAPIKey';
+      case TileMapEnum.vectorDark:
+        return 'https://maps.vietmap.vn/maps/styles/dm/style.json?apikey=$_vietmapAPIKey';
+      case TileMapEnum.rasterDefault:
+        return 'https://maps.vietmap.vn/maps/styles/tm/tiles.json?apikey=$_vietmapAPIKey';
+      case TileMapEnum.rasterLight:
+        return 'https://maps.vietmap.vn/maps/styles/lm/tiles.json?apikey=$_vietmapAPIKey';
+      case TileMapEnum.rasterDark:
+        return 'https://maps.vietmap.vn/maps/styles/dm/tiles.json?apikey=$_vietmapAPIKey';
+      default:
+        return 'https://maps.vietmap.vn/maps/styles/tm/style.json?apikey=$_vietmapAPIKey';
+    }
+  }
+
   /// Get base url for Vietmap APIs
   static String getVietmapBaseUrl() {
     if (_vietmapAPIKey.isEmpty) {
@@ -55,13 +87,25 @@ class Vietmap {
   // developers to integrate autocomplete functionality into their applications.
   // This API is designed to help users quickly find and select items from a
   // large set of options by suggesting potential matches as they type.
-  //The API is built on a machine learning model that analyzes user input and
+  // The API is built on a machine learning model that analyzes user input and
   // suggests potential matches based on the context of the search. This model
   // can be updated in real-time, allowing the API to continuously improve its
   // suggestions as more data becomes available
+  @Deprecated(
+      'Migrate VietMap API from v3 to v4. Use autocomplete instead. See https://maps.vietmap.vn/docs/map-api/autocomplete-version/autocomplete-v4')
   static Future<Either<Failure, List<VietmapAutocompleteModel>>> autocomplete(
       VietMapAutoCompleteParams params) {
     return SearchAddressUseCase(getVietmapApiRepositories()).call(params);
+  }
+
+  /// Autocomplete V4 API provides location suggestions based on user input.
+  /// This API is designed to help users quickly find and select locations
+  /// by suggesting potential matches as they type, enhancing the user experience
+  /// in location-based applications.
+  /// See more at https://maps.vietmap.vn/docs/map-api/autocomplete-version/autocomplete-v4
+  static Future<Either<Failure, List<VietmapAutocompleteModelV4>>>
+      autocompleteV4(VietmapAutocompleteParamsV4 params) {
+    return SearchAddressV4Usecase(getVietmapApiRepositories()).call(params);
   }
 
   // Updating Reverse 3.0 API is a valuable resource for developers who want to
@@ -70,9 +114,22 @@ class Vietmap {
   // and techniques, this latest version can swiftly deliver precise search
   // results for users. This API is a powerful tool that can help enhance the
   // user experience of location-based applications.
+  @Deprecated(
+      'Migrate VietMap API from v3 to v4. Use reverse instead. See https://maps.vietmap.vn/docs/map-api/reverse-version/reverse-v4')
   static Future<Either<Failure, VietmapReverseModel>> reverse(LatLng location) {
     return GetLocationFromLatLngUseCase(getVietmapApiRepositories())
         .call(LatLng(location.latitude, location.longitude));
+  }
+
+  /// Reverse V4 API is a demonstration version of the reverse geocoding service
+  /// that allows developers to test and explore location search features.
+  /// This API provides a way to convert geographic coordinates (latitude and longitude) into readable addresses and location information,
+  /// helping developers understand the capabilities before implementing the full version.
+  /// See more at https://maps.vietmap.vn/docs/map-api/reverse-version/reverse-v4.
+  static Future<Either<Failure, VietmapReverseModelV4>> reverseV4(
+      VietmapReverseParams params) {
+    return GetLocationFromLatlngV4Usecase(getVietmapApiRepositories())
+        .call(params);
   }
 
   // The Matrix API calculate many-to-many distances and times a lot more
@@ -84,8 +141,16 @@ class Vietmap {
 
   // The Place API service endpoint provides detailed information about the
   // Place found by its identifier (refid).
+  @Deprecated(
+      'Migrate VietMap API from v3 to v4. Use autocomplete instead. See https://maps.vietmap.vn/docs/map-api/autocomplete-version/autocomplete-v4')
   static Future<Either<Failure, VietmapPlaceModel>> place(String placeId) {
     return GetPlaceDetailUseCase(getVietmapApiRepositories()).call(placeId);
+  }
+
+  /// The Place API service endpoint provides detailed information about the Place found by its identifier (refid).
+  /// See more at https://maps.vietmap.vn/docs/map-api/place-v4
+  static Future<Either<Failure, VietmapPlaceModel>> placeV4(String refId) {
+    return GetPlaceDetailV4Usecase(getVietmapApiRepositories()).call(refId);
   }
 
   // A Route Maps API is a feature provided by VIETMAP that allows developers to
@@ -105,10 +170,31 @@ class Vietmap {
 
   // Updating Geocode 3.0 API is a powerful tool for developers to integrate
   // location search functionality into their applications with optimized
-  //performance. Additionally, this latest version utilizes intelligent search
-  //algorithms and methods to provide accurate and speedy search results for users.
+  // performance. Additionally, this latest version utilizes intelligent search
+  // algorithms and methods to provide accurate and speedy search results for users.
+  @Deprecated(
+      'Migrate VietMap API from v3 to v4. Use autocomplete instead. See https://maps.vietmap.vn/docs/map-api/autocomplete-version/autocomplete-v4')
   static Future<Either<Failure, List<VietmapAutocompleteModel>>> geoCode(
       VietMapAutoCompleteParams params) {
     return GeoCodeUseCase(getVietmapApiRepositories()).call(params);
+  }
+
+  /// Retrieves a list of locations based on the provided address input.
+  /// Upgraded from the Geocode v3 API, version 4 offers enhanced performance
+  /// and provides a more powerful solution for developers to integrate
+  /// location search functionality into their applications.
+  /// See more at https://maps.vietmap.vn/docs/map-api/geocode-version/geocode-v4
+  static Future<Either<Failure, List<VietmapAutocompleteModelV4>>> geoCodeV4(
+      VietmapAutocompleteParamsV4 params) {
+    return GeocodeV4Usecase(getVietmapApiRepositories()).call(params);
+  }
+
+  /// Migrate addresses between the old and new standardized formats used by VIETMAP.
+  /// This API allows users to input either an old or new address format
+  /// and receive a structured response with the corresponding converted address.
+  /// See more at https://maps.vietmap.vn/docs/migrate-address/migrate-address-docs
+  static Future<Either<Failure, VietmapMigrateAddressModel>> migrateAddress(
+      VietmapMigrateAddressParams params) {
+    return MigrateAddressUsecase(getVietmapApiRepositories()).call(params);
   }
 }