+++ lib/src/domain/repository/vietmap_api_repositories.dart
@@ -4,9 +4,15 @@ import 'package:dio/dio.dart';
 import 'package:dartz/dartz.dart';
 import 'package:flutter/foundation.dart';
 import 'package:vietmap_flutter_plugin/src/core/enums/failure_enum.dart';
+import 'package:vietmap_flutter_plugin/src/data/models/vietmap_autocomplete_model_v4.dart';
 import 'package:vietmap_flutter_plugin/src/data/models/vietmap_matrix_model.dart';
+import 'package:vietmap_flutter_plugin/src/data/models/vietmap_migrate_address_model.dart';
+import 'package:vietmap_flutter_plugin/src/data/models/vietmap_reverse_model_v4.dart';
+import 'package:vietmap_flutter_plugin/src/domain/entities/vietmap_autocomplete_params_v4.dart';
 import 'package:vietmap_flutter_plugin/src/domain/entities/vietmap_autocomplete_params.dart';
 import 'package:vietmap_flutter_plugin/src/domain/entities/vietmap_matrix_params.dart';
+import 'package:vietmap_flutter_plugin/src/domain/entities/vietmap_migrate_address_params.dart';
+import 'package:vietmap_flutter_plugin/src/domain/entities/vietmap_reverse_params.dart';
 import 'package:vietmap_flutter_plugin/src/extension/latlng_extension.dart';
 import 'package:vietmap_flutter_plugin/src/extension/string_extension.dart';
 import '../../core/failures/api_server_failure.dart';
@@ -28,6 +34,167 @@ class VietmapApiRepositories implements VietmapApiRepository {
   String apiKey = Vietmap.getVietmapAPIKey();
   VietmapApiRepositories() {
     _dio = Dio(BaseOptions(baseUrl: baseUrl));
+  }
+
+  @override
+  Future<Either<Failure, VietmapMigrateAddressModel>> migrateAddress(
+      VietmapMigrateAddressParams params) async {
+    try {
+      if (params.text.isEmpty) {
+        return Left(ExceptionFailure(
+            Exception('Địa chỉ hiện tại không được để trống')));
+      }
+      if (params.migrateType != null &&
+          params.migrateType!.value == 2 &&
+          params.focus == null) {
+        return Left(ExceptionFailure(
+            Exception('Vui lòng cung cấp tọa độ focus khi migrateType là 2')));
+      }
+      var res = await _dio.get(
+        'migrate-address/v3',
+        queryParameters: {'apikey': apiKey, ...params.toQueryParam()},
+      );
+
+      if (res.statusCode == 200) {
+        var data = VietmapMigrateAddressModel.fromJson(res.data);
+        return Right(data);
+      } else if (res.statusCode == 429) {
+        return Left(APIKeyLimitedFailure(
+            message: 'API key đã đạt giới hạn lượng truy cập hôm nay',
+            code: FailureCode.limitExceedFailure));
+      } else {
+        return const Left(ApiServerFailure('Có lỗi xảy ra'));
+      }
+    } on DioException catch (ex) {
+      if (ex.type == DioExceptionType.receiveTimeout) {
+        return Left(ApiTimeOutFailure());
+      } else {
+        return Left(ExceptionFailure(ex));
+      }
+    }
+  }
+
+  @override
+  Future<Either<Failure, VietmapPlaceModel>> getPlaceDetailV4(
+      String refId) async {
+    try {
+      if (refId.isEmpty) {
+        return Left(ExceptionFailure(Exception('RefId không được để trống')));
+      }
+      var res = await _dio
+          .get('place/v4', queryParameters: {'apikey': apiKey, 'refid': refId});
+
+      if (res.statusCode == 200) {
+        var data = VietmapPlaceModel.fromJson(res.data);
+        return Right(data);
+      } else if (res.statusCode == 429) {
+        return Left(APIKeyLimitedFailure(
+            message: 'API key đã đạt giới hạn lượng truy cập hôm nay',
+            code: FailureCode.limitExceedFailure));
+      } else {
+        return const Left(ApiServerFailure('Có lỗi xảy ra'));
+      }
+    } on DioException catch (ex) {
+      if (ex.type == DioExceptionType.receiveTimeout) {
+        return Left(ApiTimeOutFailure());
+      } else {
+        return Left(ExceptionFailure(ex));
+      }
+    }
+  }
+
+  @override
+  Future<Either<Failure, List<VietmapAutocompleteModelV4>>> geoCodeV4(
+      VietmapAutocompleteParamsV4 params) async {
+    try {
+      if (params.textSearch.isEmpty) {
+        return const Right([]);
+      }
+      final queryParams = {'apikey': apiKey, ...params.toQueryParam()};
+      var res = await _dio.get(
+        'search/v4',
+        queryParameters: queryParams,
+      );
+
+      if (res.statusCode == 200) {
+        var data = List<VietmapAutocompleteModelV4>.from(
+            res.data.map((e) => VietmapAutocompleteModelV4.fromJson(e)));
+        return Right(data);
+      } else if (res.statusCode == 429) {
+        return Left(APIKeyLimitedFailure(
+            message: 'API key đã đạt giới hạn lượng truy cập hôm nay',
+            code: FailureCode.limitExceedFailure));
+      } else {
+        return const Left(ApiServerFailure('Có lỗi xảy ra'));
+      }
+    } on DioException catch (ex) {
+      if (ex.type == DioExceptionType.receiveTimeout) {
+        return Left(ApiTimeOutFailure());
+      } else {
+        return Left(ExceptionFailure(ex));
+      }
+    }
+  }
+
+  @override
+  Future<Either<Failure, VietmapReverseModelV4>> getLocationFromLatLngV4(
+      VietmapReverseParams params) async {
+    try {
+      var res = await _dio.get('reverse/v4', queryParameters: {
+        'apikey': apiKey,
+        'lat': params.latLng.latitude,
+        'lng': params.latLng.longitude,
+        if (params.displayType != null)
+          'display_type': params.displayType!.value
+      });
+
+      if (res.statusCode == 200 && res.data.length > 0) {
+        var data = VietmapReverseModelV4.fromJson(res.data[0]);
+        return Right(data);
+      } else if (res.statusCode == 429) {
+        return Left(APIKeyLimitedFailure(
+            message: 'API key đã đạt giới hạn lượng truy cập hôm nay',
+            code: FailureCode.limitExceedFailure));
+      } else {
+        return const Left(ApiServerFailure('Có lỗi xảy ra'));
+      }
+    } on DioException catch (ex) {
+      if (ex.type == DioExceptionType.receiveTimeout) {
+        return Left(ApiTimeOutFailure());
+      } else {
+        return Left(ExceptionFailure(ex));
+      }
+    }
+  }
+
+  @override
+  Future<Either<Failure, List<VietmapAutocompleteModelV4>>> searchLocationV4(
+      VietmapAutocompleteParamsV4 params) async {
+    try {
+      final queryParams = {'apikey': apiKey, ...params.toQueryParam()};
+      var res = await _dio.get(
+        'autocomplete/v4',
+        queryParameters: queryParams,
+      );
+
+      if (res.statusCode == 200) {
+        var data = List<VietmapAutocompleteModelV4>.from(
+            res.data.map((e) => VietmapAutocompleteModelV4.fromJson(e)));
+        return Right(data);
+      } else if (res.statusCode == 429) {
+        return Left(APIKeyLimitedFailure(
+            message: 'API key đã đạt giới hạn lượng truy cập hôm nay',
+            code: FailureCode.limitExceedFailure));
+      } else {
+        return const Left(ApiServerFailure('Có lỗi xảy ra'));
+      }
+    } on DioException catch (ex) {
+      if (ex.type == DioExceptionType.receiveTimeout) {
+        return Left(ApiTimeOutFailure());
+      } else {
+        return Left(ExceptionFailure(ex));
+      }
+    }
   }
 
   @override